# Parse paths and identifiers only once