In the old implementation each variable in an expression would be parsed up to three times:

* Try to parse a path because it contains a leading double colon, or infix double colons.
* Try to parse it as path again by scanning for an identifier that contains an upper case character.
* Fall back to scanning for any identifier.

This PR turns all three steps into one, without the need for backtracking.