@Kijewski 2023-07-21
As benchmark I used `cd testing`; `find -name '*.rs' -exec touch '{}' '+'`; `time cargo b  --tests` (test out of three runs):

```
       real     user    sys
old  1.623s  10.442s  2.235s
new  1.473s  10.483s  2.272s
```

I cannot say why the real time got so much better, if the user time got worse? All CPU cores only have around 33°C, so they should all be utilized in my test without thermal throttling. I guess I should try to find a better benchmark, but I guess without #834 we cannot properly benchmark the parser. :-/ I'll try to review your PR this weekend!