@Kijewski 2023-08-02
Not a huge difference, but it's a little better:

```text
100 bytes               time:   [11.475 µs 11.594 µs 11.694 µs]
                        change: [-6.8066% -4.5471% -2.2145%] (p = 0.00 < 0.05)
                        Performance has improved.

1000 bytes              time:   [62.269 µs 62.421 µs 62.555 µs]
                        change: [-4.8183% -4.1854% -3.5724%] (p = 0.00 < 0.05)
                        Performance has improved.

10000 bytes             time:   [536.05 µs 536.83 µs 537.69 µs]
                        change: [-3.2005% -2.9935% -2.7835%] (p = 0.00 < 0.05)
                        Performance has improved.

100000 bytes            time:   [4.9944 ms 4.9996 ms 5.0053 ms]
                        change: [-2.8939% -2.7114% -2.5330%] (p = 0.00 < 0.05)
                        Performance has improved.
```